# [Feature] Block user [Fix] current chat in bold on Linux [Fix] Username is bold on Linux