--- a/project.clj
+++ b/project.clj
@@ -13,7 +13,8 @@
                  [com.cognitect/transit-cljs "0.8.248"]
                  [status-im/pluto "iteration-4-1"]
                  [mvxcvi/alphabase "1.0.0"]
-                 [rasom/cljs-react-navigation "0.1.4"]]
+                 [rasom/cljs-react-navigation "0.1.4"]
+                 [tailrecursion/cljs-priority-map "1.2.1"]]
   :plugins [[lein-cljsbuild "1.1.7"]
             [lein-re-frisk "0.5.8"]
             [lein-cljfmt "0.5.7"]
@@ -34,8 +35,8 @@
             ["do" "clean"
              ["with-profile" "prod" "cljsbuild" "once" "ios"]]
             "prod-build-desktop"     ^{:doc "Recompile code for desktop with prod profile."}
-             ["do" "clean"
-              ["with-profile" "prod" "cljsbuild" "once" "desktop"]]
+            ["do" "clean"
+             ["with-profile" "prod" "cljsbuild" "once" "desktop"]]
             "figwheel-repl"      ["with-profile" "+figwheel" "run" "-m" "clojure.main" "env/dev/run.clj"]
             "test-cljs"          ["with-profile" "test" "doo" "node" "test" "once"]
             "test-protocol"      ["with-profile" "test" "doo" "node" "protocol" "once"]
@@ -64,7 +65,7 @@
                         :repl-options {:nrepl-middleware [cider.piggieback/wrap-cljs-repl]
                                        :timeout          240000}}
              :figwheel [:dev
-                        {:dependencies [[figwheel-sidecar "0.5.16-SNAPSHOT"]
+                        {:dependencies [[figwheel-sidecar "0.5.18-SNAPSHOT"]
                                         [re-frisk-remote "0.5.5"]
                                         [re-frisk-sidecar "0.5.7"]
                                         [day8.re-frame/tracing "0.5.0"]