… src/status_im/chat/models/message.cljs
… src/status_im/chat/subs.cljs
… src/status_im/contact/core.cljs
… src/status_im/contact/db.cljs
… src/status_im/contact/subs.cljs
… src/status_im/data_store/realm/core.cljs
… src/status_im/models/wallet.cljs
… src/status_im/search/subs.cljs
… src/status_im/ui/components/chat.cljs
… src/status_im/ui/components/chat_icon/screen.cljs
… src/status_im/ui/components/chat_icon/styles.cljs
… src/status_im/ui/components/common/common.cljs
… src/status_im/ui/components/connectivity/view.cljs
… src/status_im/ui/components/desktop/tabs.cljs
… src/status_im/ui/components/icons/vector_icons.cljs
… src/status_im/ui/components/list_selection.cljs
… src/status_im/ui/components/toolbar/view.cljs
… src/status_im/ui/screens/add_new/new_chat/views.cljs
… src/status_im/ui/screens/add_new/open_dapp/views.cljs
… src/status_im/ui/screens/chat/bottom_info.cljs
… src/status_im/ui/screens/chat/input/animations/expandable.cljs
… src/status_im/ui/screens/chat/input/input.cljs
… src/status_im/ui/screens/chat/input/parameter_box.cljs
… src/status_im/ui/screens/chat/input/send_button.cljs
… src/status_im/ui/screens/chat/input/suggestions.cljs
… src/status_im/ui/screens/chat/input/validation_messages.cljs
… src/status_im/ui/screens/chat/message/message.cljs
… src/status_im/ui/screens/chat/message/options.cljs
… src/status_im/ui/screens/chat/photos.cljs
… src/status_im/ui/screens/chat/toolbar_content.cljs