fix #6033 remove logic from view
fix #1660 block user

- usernames in bold font on Linux
- prevent use of rtl char in username to impersonate
other other user random name
- current chat in bold font
- remove defview and subscriptions to use pure functions with params
- allow user to block/unblock a contact
- this will remove blocked user from chat list and
hid any message sent on any channel by blocked user

# Step to test

Plateform: desktop

sanity check
- test that chats, public chats and private chats are working correctly, responses are displayed properly
- test that users can be blocked in their profile

status: ready