@churik 2018-11-09
New issue is appeared, reproducible only in PR build.
Chat view is broken if you send several messages to chat with last message from yesterday or if you set custom date in the past in your settings (whole or part of chat area is blank)

Possible to reproduce if you have chats with last messages from yesterday or if you set custom date on your laptop to the past.

To emulate it from scratch:
- set your Mac OSx date to past (i.e. yesterday)
- create public channel
- try to send message to this channel

Expected: can see message
Actual: can see only blank screen
Video: http://take.ms/0bpgs
Logs: [Status.log](https://github.com/status-im/status-react/files/2565736/Status.log), [geth.log](https://github.com/status-im/status-react/files/2565738/geth.log)