@yenda 2018-11-09
@churik I fixed it, I think an easiest way to reproduce was to open a new public chat and then switch to already open ones, at least for me it was like this.

This PR now introduces the following glitch when replying to a chat:

![screenshot from 2018-11-09 15-44-47](https://user-images.githubusercontent.com/1181225/48269006-cfc59700-e436-11e8-92e7-bbdf71871934.png)

Personally I don't think this is blocking, and it should make fixing flat-list a priority wdyt ?