@churik 2018-11-09
@yenda 
I believe it should be fixed before merge.
Thanks for creating corresponding issue on RN-side!
Blocked by https://github.com/status-im/react-native-desktop/issues/391