@yenda 2018-11-14
@camellos yes 5000SNT by Nov 30th if I didn't add new tests by then
Regarding the length it's already cut into smaller commits, the issue was that the first commit on which it is built was blocked by a UI glitch in react-native-desktop