@churik 2018-11-14
Testing is still WIP, comment will be updated:

Mac OSx:
###  1. Last row is cut in chat list if message contains 1+ rows*
Screen:
<img width="568" alt="macos" src="https://user-images.githubusercontent.com/4557972/48496261-f23e2280-e831-11e8-9f6f-91358173e93e.png">

### 2. Space between last message and previous ones are different
<img width="568" alt="hsdf" src="https://user-images.githubusercontent.com/4557972/48496793-1817f700-e833-11e8-8243-e7739e2e0f84.png">
The space between last message and text input area is also depending on message is incoming or outgoing:
<img width="568" alt="outg" src="https://user-images.githubusercontent.com/4557972/48498642-f6b90a00-e836-11e8-8687-617e68f116a0.png">
<img width="568" alt="inc" src="https://user-images.githubusercontent.com/4557972/48498644-f751a080-e836-11e8-8dd0-59c45269d077.png">

### 3. No "Add to contacts" in 1-1 chat when someone adds you as contact
Steps:
1) log in to desktop app as user A
2) tap on Share profile code
3) log in as another user on mobile
4) start new chat > scan QR
5) send message to user A

**Expected result: **

User A can see "Add to contacts " in 1-1 chat

**Actual result: **

<img width="568" alt="add_to_contacts" src="https://user-images.githubusercontent.com/4557972/48499070-ddfd2400-e837-11e8-8b5e-9ae3ac126653.png">

### 4. Spam window doesn't block user from sending new messages 
Steps:
1) log in to desktop app as user A
2) join any public chat
3) send quickly 5+ messages in a row

**Expected result: **

Can send only 5 messages, then popup appears and user cannot send more until popup will be closed

**Actual result: **
Spam window appears, but user still can send messages 
<img width="568" alt="spam" src="https://user-images.githubusercontent.com/4557972/48499819-94add400-e839-11e8-8340-ccac8da16bdf.png">


### 5. Switching beween chats is slowed down 
Video from current PR:  https://monosnap.com/file/4esSAd79M1NTzRKJy6qdGwhZHXCqvs
Video from nightly: http://take.ms/HWVui