@yenda 2018-11-14
1. seems like an OSX specific problem, I increased the max size because on Linux the bottom of some letters was cut. I changed it back for OSX.
2. sounds like a minor issue could we bounty it ? The second part is same as in develop, it's because of the message status on the bottom right corner
3. fixed
4. it's in develop as well