@yenda 2018-11-14
5. regarding the slow down it's due to the fact that some stuff is now recalculated, I will work on optimizations later on