@chadyj 2018-11-15
Bit more context for blocking >> https://github.com/status-im/status-react/issues/6735