@churik 2018-11-15
@yenda thank you!
### 2. First part is minor, agree, but second it is not working the same on current nightly and PR.

On current nightly:
<img width="573" alt="in_outg" src="https://user-images.githubusercontent.com/4557972/48542213-df256400-e8be-11e8-8e37-4d6966ace9aa.png">

On PR build:
<img width="560" alt="cur_pr" src="https://user-images.githubusercontent.com/4557972/48548760-0fc0ca00-e8ce-11e8-8c3c-0e01cfc41270.png">
Difference is noticeable.

### 6. `Invalid arity:` 1 when creating account on IOS and Android
After creating account I've got an error:
<img width="322" alt="android_png" src="https://user-images.githubusercontent.com/4557972/48546998-f158cf80-e8c9-11e8-90f4-fe9b1f374234.png">
After restarting app and login I see the same error.
Testing on mobile is blocked.

### 7. When you tap on "Block user" on profile, then click on "Send message", you  will be redirected to empty chat without UI

Video: http://take.ms/UvhXT
Log: [Status.log](https://github.com/status-im/status-react/files/2585002/Status.log)
I you tried to send some messages to this chat, receiver will receive them, but you cannot see them (only empty screen will be shown)
<img width="568" alt="monosnap 2018-11-15 12-31-58" src="https://user-images.githubusercontent.com/4557972/48550255-8f509800-e8d2-11e8-87ca-f56d3ea464ee.png">


### 8. `TypeError: e.call is not a function` when tried to block a user who is not in my contact list
Steps:
- created new account on release build 0.9.30 (User A)
- login to desktop app (User B)
- click on "Share my contact code" (User B)
- added user B as contact (user A)
- block user  from his profile (User B)
<img width="568" alt="error_0" src="https://user-images.githubusercontent.com/4557972/48549681-be660a00-e8d0-11e8-9883-b4982b2695a9.png">
Log: [Status.log](https://github.com/status-im/status-react/files/2585040/Status.log)

### 9. Messages that are coming from blocked user are counted in unread message badge on Mac OSx

This badge is updated when messages from blocked user are coming.
<img width="28" alt="unread_0" src="https://user-images.githubusercontent.com/4557972/48550467-1aca2900-e8d3-11e8-8112-ff0206358d2b.png">

### 10. Should be optimized in terms of performance before merge.
Tests were made on the same account with same chats on different builds.

Action | CPU, % (PR) | Real memory size (PR) | CPU, % (nightly) | Real memory size (nightly) | % increase on Memory Size in PR build
-- | -- | -- | -- | -- | --
Login | 1.79 | 429 | 3 | 385.9 | 11.17
Switching between chats (10 times) | 26.52 | 1010.9 | 27 | 512 | **97.44**
Join chat with 10+ messages | 3 | 591.2 | 11 | 469.2 | 26
Send 10 messages in a row | 1.6 | 772.9 | 1.79 | 500.3 | **54.49**
Switching to Profile and back (5 times) | 27 | 444.4 | 35 | 433.5 | 2.51
Scrolling chat history | 98 | 3190 | 27 | 704.9 | **352.55**
Switching mailserver | 9 | 665.3 | 4 | 626.4 | 6.21