@yenda 2018-11-16
@churik 

I fixed 7 and 8, can you test performances again they should be better now